select result is `i32`

+++ tests/relational_tests.rs
@@ -328,7 +328,7 @@ pub async fn group_by() {
     #[derive(Debug, FromQueryResult)]
     struct SelectResult {
         name: String,
-        number_orders: Option<i64>,
+        number_orders: Option<i32>,
         total_spent: Option<Decimal>,
         min_spent: Option<Decimal>,
         max_spent: Option<Decimal>,
